# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get list of films",
                "operationId": "App\\Http\\Controllers\\FilmController::index",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Add a new film",
                "operationId": "App\\Http\\Controllers\\FilmController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "release_year": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    },
                                    "special_features": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "title",
                                    "release_year": "2000",
                                    "length": "100",
                                    "description": "description",
                                    "rating": "G",
                                    "language_id": 0,
                                    "special_features": "features",
                                    "image": "image",
                                    "created_at": "2000-01-01"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/films/{id}": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get specific film",
                "description": "Returns film data",
                "operationId": "getFilmById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Update a specific film",
                "operationId": "App\\Http\\Controllers\\FilmController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "release_year": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string"
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    },
                                    "special_features": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "title",
                                    "release_year": "2000",
                                    "length": "100",
                                    "description": "description",
                                    "rating": "G",
                                    "language_id": 0,
                                    "special_features": "features",
                                    "image": "image",
                                    "created_at": "2000-01-01"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Update a specific film",
                "operationId": "App\\Http\\Controllers\\FilmController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Deleted Film"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}